Move replicate helper into test/lib
